Files/FreeLancer Website/E-Commerce/requirements.txt
Django==5.0.2
djangorestframework==3.15.0
django-cors-headers==4.3.1
Pillow>=10.0.0
python-decouple==3.8
psycopg2-binary==2.9.9
stripe==8.4.0
django-allauth==0.60.1
django-crispy-forms==2.1
crispy-bootstrap5==0.7
django-environ==0.11.2
whitenoise==6.6.0
gunicorn==21.2.0
